users now linked to goal categories, better error handling

+++ server/controllers/goalListCtrl.ts
@@ -2,10 +2,16 @@ import { RequestHandler } from "express";
 import { isValidObjectId } from "mongoose";
 import createHttpError from "http-errors";
 import GoalList from "../models/goalList";
+import assertIsDefined from "../util/assertIsDefined";
 
 export const getAllLists: RequestHandler = async (req, res, next) => {
+  const authenticatedUserId = req.session.userId;
   try {
-    const goalLists = await GoalList.find().exec();
+    assertIsDefined(authenticatedUserId);
+
+    const goalLists = await GoalList.find({
+      userId: authenticatedUserId,
+    }).exec();
     res.status(200).json(goalLists);
   } catch (error) {
     next(error);
@@ -14,7 +20,10 @@ export const getAllLists: RequestHandler = async (req, res, next) => {
 
 export const getOneList: RequestHandler = async (req, res, next) => {
   const goalListId = req.params.id;
+  const authenticatedUserId = req.session.userId;
   try {
+    assertIsDefined(authenticatedUserId);
+
     if (!isValidObjectId(goalListId)) {
       throw createHttpError(400, "Invalid id.");
     }
@@ -23,6 +32,10 @@ export const getOneList: RequestHandler = async (req, res, next) => {
 
     if (!goalList) {
       throw createHttpError(404, "Goal not found.");
+    }
+
+    if (!goalList.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, "Unauthorized");
     }
 
     res.status(200).json(goalList);
@@ -48,12 +61,21 @@ export const createGoalList: RequestHandler<
   const desc = req.body.desc;
   const style = req.body.style; // TODO add default style
   const goals = req.body.goals;
+  const authenticatedUserId = req.session.userId;
   try {
+    assertIsDefined(authenticatedUserId);
+
     if (!title) {
       throw createHttpError(400, "Title is required.");
     }
 
-    const newGoalList = await GoalList.create({ title, desc, style, goals });
+    const newGoalList = await GoalList.create({
+      userId: authenticatedUserId,
+      title,
+      desc,
+      style,
+      goals,
+    });
 
     res.status(201).json(newGoalList);
   } catch (error) {
@@ -72,7 +94,10 @@ export const updateGoalList: RequestHandler<
   unknown
 > = async (req, res, next) => {
   const goalListId = req.params.id;
+  const authenticatedUserId = req.session.userId;
   try {
+    assertIsDefined(authenticatedUserId);
+
     if (!isValidObjectId(goalListId)) {
       throw createHttpError(400, "Invalid id.");
     }
@@ -81,6 +106,10 @@ export const updateGoalList: RequestHandler<
 
     if (!goalList) {
       throw createHttpError(404, "Goal list not found.");
+    }
+
+    if (!goalList.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, "Unauthorized");
     }
 
     goalList.title = req.body.title || goalList.title;
@@ -107,7 +136,10 @@ export const addGoalToList: RequestHandler<
 > = async (req, res, next) => {
   const goalListId = req.params.id;
   const goal = req.body.goal;
+  const authenticatedUserId = req.session.userId;
   try {
+    assertIsDefined(authenticatedUserId);
+
     if (!goal || !goal.text) {
       throw createHttpError(400, "Goal cannot be empty.");
     }
@@ -120,6 +152,10 @@ export const addGoalToList: RequestHandler<
 
     if (!goalList) {
       throw createHttpError(404, "Goal list not found.");
+    }
+
+    if (!goalList.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, "Unauthorized");
     }
 
     const newGoal = {
@@ -143,7 +179,10 @@ export const removeGoalFromList: RequestHandler<
 > = async (req, res, next) => {
   const goalListId = req.params.id;
   const goal = req.body.goal;
+  const authenticatedUserId = req.session.userId;
   try {
+    assertIsDefined(authenticatedUserId);
+
     if (!goal || !goal.text) {
       throw createHttpError(400, "No goal to delete.");
     }
@@ -158,6 +197,10 @@ export const removeGoalFromList: RequestHandler<
       throw createHttpError(404, "Goal list not found.");
     }
 
+    if (!goalList.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, "Unauthorized");
+    }
+
     goalList.goals.pull({ text: goal.text });
     await goalList.save();
     res.status(200).json(goalList);
@@ -168,7 +211,10 @@ export const removeGoalFromList: RequestHandler<
 
 export const deleteGoalList: RequestHandler = async (req, res, next) => {
   const goalListId = req.params.id;
+  const authenticatedUserId = req.session.userId;
   try {
+    assertIsDefined(authenticatedUserId);
+
     if (!isValidObjectId(goalListId)) {
       throw createHttpError(400, "Invalid id.");
     }
@@ -177,6 +223,10 @@ export const deleteGoalList: RequestHandler = async (req, res, next) => {
 
     if (!goalList) {
       throw createHttpError(404, "Goal list not found.");
+    }
+
+    if (!goalList.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, "Unauthorized");
     }
 
     await GoalList.findByIdAndDelete(goalListId);
